# docs/swagger.yaml
basePath: /
definitions:
...
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  stores.CreateStoreRequest:
    properties:
      address:
        type: string
      name:
        type: string
    required:
      - address
      - name
    type: object
  stores.StoreResponse:
    properties:
      address:
        type: string
      name:
        type: string
      uuid:
        type: string
    type: object
  stores.UpdateStoreRequest:
    properties:
      address:
        type: string
      name:
        type: string
    required:
      - address
      - name
    type: object
host: localhost:8080
info:
  contact: { }
  description: Это REST API для сервиса сети книжных магазинов.
  title: Bookstores API
  version: "1.0"
paths:
...
  /stores:
    get:
      description: Возвращает список всех действующих магазинов.
      produces:
        - application/json
      responses:
        "200":
          description: Список действующих магазинов
          schema:
            items:
              $ref: '#/definitions/stores.StoreResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить список магазинов
      tags:
        - stores
    post:
      consumes:
        - application/json
      description: Создаёт новую запись о магазине в базе данных.
      parameters:
        - description: Данные для создания магазина
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/stores.CreateStoreRequest'
      produces:
        - application/json
      responses:
        "201":
          description: Магазин успешно создан
          schema:
            $ref: '#/definitions/stores.StoreResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создать новый магазин
      tags:
        - stores
  /stores/{storeUUID}:
    delete:
      description: Выполняет мягкое удаление магазина
      parameters:
        - description: UUID магазина
          in: path
          name: storeUUID
          required: true
          type: string
      responses:
        "204":
          description: Магазин удалён (деактивирован)
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Удалить магазин из доступных
      tags:
        - stores
    get:
      description: Возвращает детальную информацию о магазине по его UUID.
      parameters:
        - description: UUID магазина
          in: path
          name: storeUUID
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Инфо о найденном магазине
          schema:
            $ref: '#/definitions/stores.StoreResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Искомый магазин отсутствует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить информацию об одном магазине
      tags:
        - stores
    put:
      consumes:
        - application/json
      description: Обнавляет запись о магазине в базе данных.
      parameters:
        - description: UUID магазина
          in: path
          name: storeUUID
          required: true
          type: string
        - description: Данные для обновления информации о магазине
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/stores.UpdateStoreRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Обновлённое инфо об обновлённом магазине
          schema:
            $ref: '#/definitions/stores.StoreResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновить информацию о магазине
      tags:
        - stores
swagger: "2.0"